fix clang warnings about GNU stuff

+++ src/shared/dlt_offline_trace.c
@@ -195,7 +195,7 @@ DltReturnValue dlt_offline_trace_create_new_file(DltOfflineTrace *trace)
                        DLT_OFFLINETRACE_FILENAME_TIMESTAMP_DELI, timestamp,
                        DLT_OFFLINETRACE_FILENAME_EXT);
 
-        if ((ret < 0) || (ret >= (int)sizeof(trace->filename))) {
+        if ((ret < 0) || ((size_t)ret >= (int)sizeof(trace->filename))) {
             printf("dlt_offlinetrace filename cannot be concatenated\n");
             return DLT_RETURN_ERROR;
         }
@@ -203,7 +203,7 @@ DltReturnValue dlt_offline_trace_create_new_file(DltOfflineTrace *trace)
         ret = snprintf(file_path, sizeof(file_path), "%s/%s",
                        trace->directory, trace->filename);
 
-        if ((ret < 0) || (ret >= (int)sizeof(file_path))) {
+        if ((ret < 0) || ((size_t)ret >= (int)sizeof(file_path))) {
             printf("dlt_offlinetrace file path cannot be concatenated\n");
             return DLT_RETURN_ERROR;
         }
